requirements.txt
django
crispy-bootstrap5
django-ckeditor
django-crispy-forms
django-debug-toolbar
Pillow
pytest-djangoasgiref==3.5.2
attrs==22.1.0
crispy-bootstrap5==0.6
cycler==0.11.0
Django==4.1.1
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-debug-toolbar==3.6.0
django-js-asset==2.0.0
fonttools==4.37.1
iniconfig==1.1.1
joblib==1.1.0
kiwisolver==1.4.4
matplotlib==3.5.3
numpy==1.23.3
packaging==21.3
pandas==1.4.4
Pillow==9.2.0
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.3
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.2.1
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.0
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.1.0
tomli==2.0.1
